Git issue resolution
