+++ cwms-data-api/src/main/java/cwms/cda/data/dao/LocationLevelsDaoImpl.java
@@ -643,7 +643,7 @@ public class LocationLevelsDaoImpl extends JooqDao<LocationLevel> implements Loc
         TimeSeries timeSeries = new TimeSeries(null, size, size, timeSeriesId,
                 officeId, firstValueTime, lastValueTime, levelUnits,
                 java.time.Duration.ofSeconds(interval.getSeconds()),
-                null, null, locationTimeZone.getId(), null, VersionType.UNVERSIONED.getValue());
+                null, null, locationTimeZone.getId(), null, VersionType.UNVERSIONED);
         for (ZTSV_TYPE tsv : locLvlValues) {
             Timestamp dateTime = tsv.getDATE_TIME();
             Double value = tsv.getVALUE();
